refactor: split middleware.rs.
